feat(sidebar): move worktree filter out of search bar into header
Filter dropdown now sits next to View options in the Workspaces header,
leaving the search bar with just the query input.
